rest api add login

+++ src/controllers/notes.controllers.js
@@ -2,7 +2,8 @@ import Note from "../models/notes.models";
 
 export const getNotes = async (req, res) => {
     try {
-        const notes = await Note.find();
+        const userId = req.user.id
+        const notes = await Note.find({userId});
         res.json(notes)
     } catch (error) {
         return res.status(400).json({
@@ -13,9 +14,10 @@ export const getNotes = async (req, res) => {
 }
 
 export const createNote = async (req, res) => {
-    const{title, description} = req.body
     try {
-        const note = new Note({title, description})
+        req.body.userId = req.user.id
+        const{title, description, userId} = req.body
+        const note = new Note({title, description, userId})
         const save = await note.save()
         res.status(201).json(save)
     } catch (error) {
@@ -29,6 +31,11 @@ export const createNote = async (req, res) => {
 export const getNote = async (req, res) => {
     try {
         const note = await Note.findById(req.params.id)
+
+        if(!(req.user.id==note.userId)){
+            return res.status(401).json({message: 'Unauthorize'})
+
+        }
         res.status(200).json(note)
     } catch (error) {
         return res.status(400).json({
@@ -41,6 +48,10 @@ export const getNote = async (req, res) => {
 
 export const updateNote = async (req, res) => {
     try {
+        const note_id = await Note.findById(req.params.id)
+        if(!(req.user.id==note_id.userId)){
+            return res.status(401).json({message: 'Unauthorize'})
+        }
         const note = await Note.findByIdAndUpdate(req.params.id, req.body, {new: true})
         res.status(200).json(note)
     } catch (error) {
@@ -53,8 +64,12 @@ export const updateNote = async (req, res) => {
 
 export const deleteNote = async (req, res) => {
     try {
+        const note_id = await Note.findById(req.params.id)
+        if(!(req.user.id==note_id.userId)){
+            return res.status(401).json({message: 'Unauthorize'})
+        }
         const note = await Note.findByIdAndDelete(req.params.id)
-        res.status(200).json(note) 
+        res.status(200).json(note)
     } catch (error) {
         return res.status(400).json({
             message: 'Error delete note',
